% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_website.R
\name{create_website}
\alias{create_website}
\title{Create the templates for a condensr website}
\usage{
create_website(dir, title)
}
\arguments{
\item{dir}{Target directory for the website.}

\item{title}{Title of the website.}
}
\description{
Creates the necessary directories, high level files and structures for a
condensr website.
}
\details{
Under the hood, this function is creating a distill website (using
\code{distill:::do_create_website}), and then overwriting some files with
\code{condensr} specifics. These specific files are:
\itemize{
\item _site.yml
\item index.Rmd
staff_list.Rmd
staff_member.Rmd
}
}
\examples{
create_website(dir = tempdir(), "test_site")
}
