% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_markdown_for_member.R
\name{generate_markdown_for_member}
\alias{generate_markdown_for_member}
\title{Generate markdown for staff member}
\usage{
generate_markdown_for_member(member)
}
\arguments{
\item{member}{An object of class \code{staff_member}.}
}
\description{
Generates the image, short description and linking-out buttons for use in
a list of staff belonging to a group.
}
