% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_markdown_for_projects.R
\name{generate_markdown_for_projects}
\alias{generate_markdown_for_projects}
\title{Generate markdown for projects}
\usage{
generate_markdown_for_projects(project)
}
\arguments{
\item{project}{A \code{condensr} project object. If a link is available (i.e.
\code{!is.na(project$link)}), then a hyperlink will be inserted.}
}
\value{
No return value, function will output text directly into markdown
document using \code{cat}.
}
\description{
Generate markdown block for \code{codensr} projects.
}
\examples{

example_project <- project(
       id = "joes-model",
       name = "Joe's Blogging Model",
       related_staff = "joe-bloggs"
   )
generate_markdown_for_projects(example_project)

}
