% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_markdown_for_publications.R
\name{generate_markdown_for_publications}
\alias{generate_markdown_for_publications}
\title{Generate markdown for publications}
\usage{
generate_markdown_for_publications(publication)
}
\arguments{
\item{publication}{A \code{condensr} publication object. If a link is available (i.e.
\code{!is.na(publication$link)}), then a hyperlink will be inserted.}
}
\description{
Generate markdown block for \code{codensr} publications.
}
