% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_template}
\alias{render_template}
\title{Copy Rmd template to target path}
\usage{
render_template(
  file,
  template_folder,
  target_path,
  output_name = NA,
  data = list()
)
}
\arguments{
\item{file}{File name to copy.}

\item{template_folder}{Folder name of file.}

\item{target_path}{Path to copy to.}

\item{output_name}{Name of copies file. If \code{NA}, name will be the same as
input file name.}

\item{data}{List of additional metadata. Must have at least element \code{name}
and \code{title}.}
}
\description{
Copies a file from the \code{rmarkdown/templates} folder to a target path
}
