% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_list_by_date.R
\name{sort_list_by_date}
\alias{sort_list_by_date}
\title{Sort a list by the list element \code{date}.}
\usage{
sort_list_by_date(list_to_sort)
}
\arguments{
\item{list_to_sort}{List to sort. Each element of the list must have column \code{date}.}
}
\value{
The sorted list.
}
\description{
Sorts a list for the column \code{date}.
}
