% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_website.R
\name{create_website}
\alias{create_website}
\title{Create the templates for a condensr website}
\usage{
create_website(dir, title)
}
\arguments{
\item{dir}{Target directory for the website.}

\item{title}{Title of the website.}
}
\value{
No return value, called to create the skeleton structure of the website.
}
\description{
Creates the necessary directories, high level files and structures for a
condensr website.
}
\details{
Under the hood, this function is creating a distill website (using
\code{distill:::do_create_website}), and then overwriting some files with
\code{condensr} specifics. These specific files are:
\itemize{
\item _site.yml - for site specific configurations
\item index.qmd - the home page of the website
\item staff_list.R - an example setup the \code{condensr} staff, project and publication lists
\item staff/index.qmd - the template page that will list staff members
\item staff/joe-bloggs/index.qmd - an example page for the hypothetical staff member
}
}
\examples{
create_website(dir = tempdir(), "test_site")
}
