% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_markdown_for_member.R
\name{generate_markdown_for_member}
\alias{generate_markdown_for_member}
\title{Generate markdown for staff member}
\usage{
generate_markdown_for_member(member, extension = "png")
}
\arguments{
\item{member}{An object of class \code{staff_member}.}

\item{extension}{Extension of the image file. Don't include the .}
}
\value{
No return value, function will output text directly into markdown
document using \code{cat}.
}
\description{
Generates the image, short description and linking-out buttons for use in
a list of staff belonging to a group.
}
\examples{
example_staff_member <- staff_member(
    id = "example-id",
    name = "Example Name",
    description = "Example description for example staff member.",
    external_link = "http://www.example.com/",
    internal_link = FALSE
)
generate_markdown_for_member(example_staff_member)

}
